requirements.txt
pandas>=2.0.0
openpyxl>=3.0.0
pyarrow>=10.0.0

# DICOM processing
SimpleITK>=2.5.0
pydicom>=3.0.0
pydicom-seg>=0.4.0
highdicom>=0.27.0
pillow>=12.0.0
numpy>=1.26.0,<2.0.0
tqdm>=4.67.0

# 3D mesh processing
trimesh>=4.0.0
scipy

# Visualization
matplotlib>=3.7.0
pillow>=12.0.0

# Logging & Config
loguru>=0.7.0
PyYAML>=6.0.0

# ML/DL
torch>=2.0.0
torchvision>=0.15.0
aim>=3.17.0

wandb
weave
loguru